Split _internal.py into directory to fix circular import issue
